address review comments from Dr. Ingrid
